feat: add server route for getting simplexes
