+++ src/common/types/dto/get-solution.dto.ts
@@ -1,0 +1,6 @@
+export interface GetSolutionRequestDto {
+  expression: string;
+}
+export interface GetSolutionResponseDto {
+  simplexes: number[][][];
+}
