+++ src/server/routes/tree.routes.ts
@@ -1,9 +1,10 @@
 import { Router } from 'express';
-import { treeController } from '../controllers/tree.controller';
+import { mainController } from '../controllers/main.controller';
 
 const router = Router();
 
-router.get('/create-tree', treeController.createTree);
-router.get('/get-graph', treeController.getGraph);
+router.get('/create-tree', mainController.createTree);
+router.get('/get-graph', mainController.getGraph);
+router.get('/get-solution', mainController.getSolution);
 
 export const treeRouter = router;
